Codexd

+++ sql/postgres-60day/day21_distribution_functions.sql
@@ -14,7 +14,7 @@ WITH cust_rev AS (
 SELECT customer_id,
        revenue,
        NTILE(4) OVER (ORDER BY revenue DESC) AS revenue_quartile,
-       ROUND(PERCENT_RANK() OVER (ORDER BY revenue),4) AS pct_rank
+       ROUND((PERCENT_RANK() OVER (ORDER BY revenue))::numeric, 4) AS pct_rank
 FROM cust_rev
 ORDER BY revenue DESC
 LIMIT 50;
